feat(api): implement admin abuse
